requirements.txt
timm==0.4.12
einops==0.6.0
numpy==1.23.5
tqdm==4.66.1
opencv-python==4.8.1.78
pillow==10.2.0
setproctitle==1.3.3
matplotlib==3.7.5
